Move socket to linux
To better describe the package, rename socket to linux. This is because
the package contained more than just socket operations.
