+++ lxd/seccomp/seccomp.go
@@ -29,10 +29,10 @@ import (
 	"github.com/lxc/lxd/shared"
 	"github.com/lxc/lxd/shared/api"
 	"github.com/lxc/lxd/shared/idmap"
+	"github.com/lxc/lxd/shared/linux"
 	"github.com/lxc/lxd/shared/logger"
 	"github.com/lxc/lxd/shared/netutils"
 	"github.com/lxc/lxd/shared/osarch"
-	"github.com/lxc/lxd/shared/socket"
 )
 
 /*
@@ -721,7 +721,7 @@ func InstanceNeedsIntercept(s *state.State, c Instance) (bool, error) {
 // MakePidFd prepares a pidfd to inherit for the init process of the container.
 func MakePidFd(pid int, s *state.State) (int, *os.File) {
 	if s.OS.PidFds {
-		pidFdFile, err := socket.PidFdOpen(pid, 0)
+		pidFdFile, err := linux.PidFdOpen(pid, 0)
 		if err != nil {
 			return -1, nil
 		}
@@ -2102,7 +2102,7 @@ func (s *Server) HandleMountSyscall(c Instance, siov *Iovec) int {
 	args.data = C.GoString(&mntData[0])
 	ctx["data"] = args.data
 
-	err := socket.PidfdSendSignal(int(pidFd.Fd()), 0, 0)
+	err := linux.PidfdSendSignal(int(pidFd.Fd()), 0, 0)
 	if err != nil {
 		ctx["err"] = fmt.Sprintf("Failed to send signal to target process for of mount syscall: %s", err)
 		ctx["syscall_continue"] = "true"
