+++ shared/linux/socket_linux_notcgo.go
@@ -1,5 +1,5 @@
 //go:build linux && !cgo
 
-package socket
+package linux
 
 const ABSTRACT_UNIX_SOCK_LEN int = 107
